# Remove `Dict{Symbol, Any}` for type stability in `FunctionOperator`