The current implementation of `FunctionOperator` requires a `Dict{Symbol,Any}` to handle the keyword arguments. This leads to type instabilities and performance decreases, as in the following simple code:

```julia
using LinearAlgebra
using BenchmarkTools
using SciMLOperators

T = ComplexF64
N = 10

u = rand(T, N)
du = similar(u)
p = (A=rand(T, N, N),)

op = FunctionOperator((du, u, p, t) -> mul!(du, p.A, u), u, p = p, T=T, isinplace=Val(true), outofplace=Val(true), has_mul5=Val(true), isconstant=true)

op(du, u, p, 0.1)

@benchmark $op($du, $u, $p, 0.1)
```

### master branch
```julia-repl
BenchmarkTools.Trial: 10000 samples with 53 evaluations.
 Range (min … max):  864.434 ns … 104.262 μs  ┊ GC (min … max): 0.00% … 98.52%
 Time  (median):     947.340 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   996.750 ns ±   1.294 μs  ┊ GC (mean ± σ):  2.57% ±  2.17%

           ▃█                                                    
  ▂▂▃▅▄▃▃▃▅███▄▃▃▃▃▃▃▃▃▄▅▄▄▃▂▂▂▂▂▂▂▂▂▂▂▁▂▁▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂ ▃
  864 ns           Histogram: frequency by time         1.34 μs <

 Memory estimate: 560 bytes, allocs estimate: 7.
```

### This PR branch

```julia-repl
BenchmarkTools.Trial: 10000 samples with 989 evaluations.
 Range (min … max):  46.024 ns … 64.056 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     48.916 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   48.858 ns ±  1.040 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▃▄▄▂  ▁▁▁▂▂▂▂▁    ▁▁▁▁▁▁▁▁▁ ▁▂▃▆▇█▇▆▅▄▅▅▅▅▅▅▄▄▄▁▁▁▁▁▁▂▂▁▁  ▃
  ▇██████████████████████████████████████████████████████████ █
  46 ns        Histogram: log(frequency) by time      50.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

---

To manage to fix this issue, I used `NamedTuple` instead of `Dict{Symbol, Any}`. For this reason, the additional `kwargs` are required when `accepted_kwargs` is provided, in order to initialize the types inside the `NamedTuple`. In principle, the `accepted_kwargs` can be directly obtained from `kwargs` keys, but I left it for consistency with the other functions.

Furthermore, I've extended the methods of the `get_filtered_kwargs`, also supporting `Val` types for `accepted_kwargs`. In this way, the filtering is performed at compile time, otherwise the type of the `NamedTuple` cannot be inferred. Thus, all the other functions supporting `accepted_kwargs` would also support `Val` types, improving performance.