@albertomercurio 2024-10-20
I've also made the `mul!` function for the `IdentityOperator`, because the following example was allocating

```julia
T = Float64
N = 10

u = rand(T, N)
du = similar(u)

A = MatrixOperator(rand(T, N, N))

op = T(0.1) * A + T(0.2) * IdentityOperator(N)
# op = T(0.2) * IdentityOperator(N)
# op = T(0.1) * A

@benchmark mul!($du, $op, $u)
```

### Master branch

```julia-repl
BenchmarkTools.Trial: 10000 samples with 982 evaluations.
 Range (min … max):  57.484 ns …  14.506 μs  ┊ GC (min … max): 0.00% … 99.22%
 Time  (median):     67.768 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   79.566 ns ± 198.549 ns  ┊ GC (mean ± σ):  6.26% ±  2.77%

           ▁▃▅█▇▂                                               
  ▂▁▂▂▃▃▅▇███████▆▄▃▃▂▂▂▂▃▂▂▃▄▅▅▄▃▃▂▂▂▂▂▃▃▄▄▄▄▅▇██▇▄▃▃▂▂▂▂▂▂▂▂ ▃
  57.5 ns         Histogram: frequency by time          101 ns <

 Memory estimate: 112 bytes, allocs estimate: 3.
```

### This PR branch

```julia-repl
BenchmarkTools.Trial: 10000 samples with 994 evaluations.
 Range (min … max):  31.203 ns … 42.741 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     33.894 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   33.676 ns ±  0.924 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                  ▁▃▅█▇▄▂▁                     
  ▁▁▁▂▁▂▂▂▃▃▅▅▅▅▅▄▄▃▂▂▂▂▂▂▂▄▄▅▅▅▇▇████████▆▅▄▃▃▃▃▃▃▂▂▂▂▂▂▁▁▁▁ ▃
  31.2 ns         Histogram: frequency by time        35.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```