@ChrisRackauckas 2024-10-21
this repo needs quite a bit of work in order to v1.0. Seems like you're starting to get into it. If you're interested in helping, maybe we can find a time to chat. The core issue is that a lot of the operators do `f(u,p,t)*u` when it needs to generally be `f(u,p,t)*v`