@albertomercurio 2024-10-21
Yes, I'm starting to use it due to its dependence of my package QuantumToolbox.jl. I don't think I have so much time for directly helping this repo, but for sure I will make other PRs when I will need it. Btw, I'm happy to have a chat and discuss about this.